change build file

+++ build.sh
@@ -3,9 +3,10 @@ set -e
 # build document
 
 MODE=${RUNNING_MODE}
-
+CLEAN_ALL_FILES_FLAG="false"
 build_user_portal() {
   set -x
+  echo "start build frontend files ..."
   outputPath="static/bundles"
   # rm -rf dist
   rm -rf ${outputPath}
@@ -22,18 +23,25 @@ build_user_portal() {
 
 build_user_portal_backend() {
   set -x
+  echo "start build backtend files ..."
   outputPath="dist"
   tmp="user_portal_tmp"
   node_modules/.bin/webpack --config webpack.config.backend.js
-  mkdir ${tmp}
-  # copy files to tmp dir
-  cp ${outputPath}/app.js ${tmp}/app.js
-  cp -rf static ${tmp}/
-  cp index.html ${tmp}/index.html
-  # rm all source files
-  ls | grep -v ${tmp} | grep -v node_modules | xargs rm -rf
-  mv ${tmp}/* ./
-  rm -rf ${tmp}
+
+  # 如果在执行脚本时传递 '--clean=all' 参数，构建完成后会删除源文件
+  if [ "${CLEAN_ALL_FILES_FLAG}" = "true" ] then
+      mkdir -p ${tmp}
+      # copy files to tmp dir
+      cp ${outputPath}/app.js ${tmp}/app.js
+      cp -rf static ${tmp}/
+      cp index.html ${tmp}/index.html
+      # rm all source files
+      ls | grep -v ${tmp} | grep -v node_modules | xargs rm -rf
+      mv ${tmp}/* ./
+      rm -rf ${tmp}
+  else
+      cp ${outputPath}/app.js ./backend.js
+  fi
 
   set +x
 }
@@ -49,10 +57,14 @@ else
   echo "start build ${project}"
   echo "node_env: ${NODE_ENV}"
   echo "running_mode: ${MODE}"
+  # build frontend files
   build_user_portal
-  # 只有在私有云，而且在执行脚本时传递 '--clean=all' 参数
-  # 才会构建后端代码，且在构建完成后会删除源文件
-  if [ "$1" = "--clean=all" ] && [ "${MODE}" = "enterprise" ]; then
+  # build backend files
+  if [ "$2" = "--clean=all" ] then
+      CLEAN_ALL_FILES_FLAG="true"
+  fi
+  # 只有在私有云，而且在执行脚本时传递 '--build=all' 参数，才会构建后端代码
+  if [ "${MODE}" = "enterprise" ] && [ "$1" = "--build=all" ]; then
       echo "***********************************"
       echo "* will build backend              *"
       echo "* will delete all source files ...*"
